requirements.txt
PyQt5>=5.15.0
pyserial>=3.5 